ai_mqtt_langchain/requirements.txt
python-dotenv==1.0.0
paho-mqtt==2.2.1
langchain==0.0.267
openai==0.27.8
RPi.GPIO==0.7.1; platform_system=="Linux" and platform_machine.startswith("arm")
influxdb-client==1.36.1
Flask==2.3.2
Flask-SocketIO==5.3.4
Flask-Cors==4.0.0
PyJWT==2.7.0
Werkzeug==2.3.6
requests==2.31.0
schedule==1.2.0
python-socketio==5.8.0
eventlet==0.33.3
gTTS==2.3.2; platform_system=="Linux" and platform_machine.startswith("arm")
gunicorn==21.2.0
pytest==7.4.0
pytest-cov==4.1.0 